Cxy621

+++ src/main/java/com/sast/atSast/mapper/TeamMemberMapper.java
@@ -10,7 +10,9 @@ import java.util.List;
 @Repository
 @Mapper
 public interface TeamMemberMapper {
-    List<TeamMember> getTeams(@Param("contestId")Integer contestId);
-    void updateTeam(@Param("teamId")Long teamId,@Param("score")Integer score,@Param("result")String result);
-    void insertTeam(@Param("teamMember")TeamMember teamMember);
+    List<TeamMember> getTeams(@Param("contestId") Long contestId);
+
+    void updateTeam(@Param("teamId") Long teamId, @Param("score") Integer score, @Param("result") String result);
+
+    void insertTeam(@Param("teamMember") TeamMember teamMember);
 }
